Splited check of false positives into two scripts so assertions can be manually modified

+++ scripts/check_false_positives.sh
@@ -34,15 +34,21 @@ fi
 IFS=';' #setting ; as delimiter
 while read -ra line; do
     echo -e "\n\n\n${YELLOW}Checking false positives for ${line[1]}${NORMAL}\n"
+    
+    assertions_file="assertions/${line[0]}/$1/${line[0]}.java"
+    if [[ ! -f "$assertions_file" ]]; then
+        echo -e "\n${RED}there are no assertions for ${line[0]} with $1${NORMAL}\n"
+        continue
+    fi
+    
     package_as_path="$( echo "${line[1]}" | tr  '.' '/'  )"
-    ./scripts/utils/insert_assertion.sh subjects/"${line[0]}"/src/main/java/$package_as_path/"${line[2]}".java "${line[2]}" "${line[3]}" "${line[0]}" $1
     
     output_dir=OASIs_results/"${line[0]}"/$1
     log=""
     create_results_folder
     add_libraries "${line[0]}"
+    
+    cp "$assertions_file" "subjects/${line[0]}/src/main/java/$package_as_path/${line[2]}.java"
     bash scripts/utils/OASIs_mod.sh "${line[1]}.${line[2]}" "subjects/${line[0]}/src/main/java" "${line[3]}" "$libraries" > $log
-    #save the modified class into results and restore the original
-    cp "subjects/${line[0]}/src/main/java/$package_as_path/${line[2]}.java" "$output_dir/${line[2]}.java"
     cp "original_subject_classes/${line[0]}/${line[2]}.java" "subjects/${line[0]}/src/main/java/$package_as_path/${line[2]}.java"
 done <$file
